requirements.txt
streamlit
PyPDF2
langchain
langchain-text-splitters
langchain-openai
langchain-community
faiss-cpu
python-dotenv
tiktoken
